chore: replace toolkit-in-action with _page
