+++ packages/button-theme/stories/1_introduction-pages/5_Toolkit_in_Action/_Page.stories.mdx
@@ -8,5 +8,7 @@ import * as stories from './_Page.stories.tsx';
 Here's an example of how the components in this toolkit can come together to make a service.
 
 <Canvas>
-  <Story story={stories.Default} name="Demo" />
+  <Story story={stories.JS} name="JS" />
+  <Story story={stories.CSS} name="CSS" />
+  <Story story={stories.HTML} name="HTML" />
 </Canvas>
